Separate database recovery orchestration
Move full-resync task ownership, retry/state-machine execution, and recovery transport helpers into a focused runtime module, with shared sync-result policy isolated for recovery and E2EE hooks.
